requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
Pillow>=10.0.0
aiohttp>=3.9.0
tipo-kgen>=0.2.0
transformers>=4.44.0
torch>=2.2.0
